# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: ToDo List System API
  version: 0.0.1
servers:
  - url: http://localhost:8080
...
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Retrieve all tasks.
      description: Retrieves a list of all Task objects.
      responses:
        200:
          description: List of Task objects was successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Creates a new Task object.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        201:
          description: Task object was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        400:
          description: Task object contains invalid data.
  /tasks/{taskId}:

    get:
      tags:
        - Tasks
      summary: Retrieve a specific task.
      description: Retrieves a Task object by its unique ID.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Task object was successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        404:
          description: Task object not found.

    put:
      tags:
        - Tasks
      summary: Update a task
      description: Updates a specific Task object.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        200:
          description: Task object was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        404:
          description: Task object not found.

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Deletes a Task object by its unique ID.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Task object was successfully deleted.
        404:
          description: Task object not found.
  /categories:
    get:
      tags:
        - Categories
      summary: Retrieve all categories.
      responses:
        200:
          description: Retrieves a list of all Category objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags:
        - Categories
      summary: Create a new Category.
      description: Creates a new Category object.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        201:
          description: Category object was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Category object contains invalid data.
  /categories/{categoryId}:
    get:
      tags:
        - Categories
      summary: Retrieve a specific Category.
      description: Retrieves a Category object by its unique ID.
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Category object was successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        404:
          description: Category object not found.

    put:
      tags:
        - Categories
      summary: Update a Category.
      description: Updates a specific Category object.
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        200:
          description: Category object was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        404:
          description: Category object not found.

    delete:
      tags:
        - Categories
      summary: Delete a Category.
      description: Deletes a Category object by its unique ID.
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Category object was successfully deleted.
        404:
          description: Category object not found.
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid

        user_id:
          type: string
          format: uuid

        title:
          type: string
          maxLength: 100

        description:
          type: string
          nullable: true

        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
          default: PENDING

        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
          default: MEDIUM

        due_date:
          type: string
          format: date-time
          nullable: true

        created_at:
          type: string
          format: date-time
          default: now()
    TaskInput:
      type: object
      properties:
        user_id:
          type: string
          format: uuid

        title:
          type: string
          maxLength: 100

        description:
          type: string
          nullable: true

        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
          default: PENDING

        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
          default: MEDIUM

        due_date:
          type: string
          format: date-time
          nullable: true
    Category:
      type: object
      required:
        - name
        - description
      properties:
        id:
          type: string
          format: uuid

        name:
          type: string

        description:
          type: string

        user_id:
          type: string
          format: uuid
    CategoryInput:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string

        description:
          type: string

        user_id:
          type: string
          format: uuid
...